+++ CHANGELOG.md
@@ -4,13 +4,17 @@
 
 ### New features
 
-- Option to change font size in recipe list
+- Option to change font size in recipe list (by default 'Large')
 - Options to select modifier keys (shift, ctrl) while selecting recipes, cetegories and tabs (by default all 'None')
+- Option to show mod name of recipe (if not 'Base', by default 'false')
+- Added new special search character (&) - searches by mod name of recipe
+- Added new token modifier in search (~) - negates this token
 
 ### Improvements
 
 - Config checking: If something new added to config (e.g options), it will autoupdate with default values
 - Moved tab closing context menu a bit higher
+- Updated search helper info
 
 ## 1.5.5 (17.04.2022)